Mobile website design optimization using Hypothesis Testing
An online bookstore is looking to optimize its website design to improve user engagement and ultimately increase book purchases.
The website currently offers two themes for its users: “Light Theme” and “Dark Theme.” The bookstore’s data science team wants to conduct an A/B testing experiment to determine which theme leads to better user engagement and higher conversion rates for book purchases.

The data collected by the bookstore contains user interactions and engagement metrics for both the Light Theme and Dark Theme. The dataset includes the following key features:

Theme: dark or light

Click Through Rate: The proportion of the users who click on links or buttons on the website.
Conversion Rate: The percentage of users who signed up on the platform after visiting for the first time.
Bounce Rate: The percentage of users who leave the website without further interaction after visiting a single page.
Scroll Depth: The depth to which users scroll through the website pages.
Age: The age of the user.
Location: The location of the user.
Session Duration: The duration of the user’s session on the website.
Purchases: Whether the user purchased the book (Yes/No).
Added_to_Cart: Whether the user added books to the cart (Yes/No).
Your task is to identify which theme, Light Theme or Dark Theme, yields better user engagement, purchases and conversion rates. You need to determine if there is a statistically significant difference in the key metrics between the two themes.

SITUATION: An online bookstore is faced with making a business decision concerning its web design optimization. The web design must be able to improve user engagement and drive sales.
TASK : To identify which theme Light or Dark will solve the problem
ACTION: To conduct an A/B Testing of to determine which theme based on the features highlighted above will encourage book purchase.
RESULT: